+++ CHANGELOG.md
@@ -3,6 +3,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [1.1.0] - 2021-07-07
+### Added
+- TimePointCalculateVelocityToHitTarget, Get a vector3 from a point in time from the CalculatedVelocityToHitTarget
+- Documentation for more complex functions
+- Readme example list
+- Editor test script to test the functions
+- SnapToGrid
+- SnapValue
+- PositionInCircle
+- EnlargeArrayValues
+- MirrorVector2
+- RotateVector2
+- ShortArray
+- ValueDiff
+- ShuffleList
+### Changed
+- Renamed "CalculateVelocityNeededToHitTarget" to "CalculateVelocityToHitTarget"
+
 ## [1.0.0] - 2021-07-07
 ### Added
 - First public release